Extract EmployeeDetailModal
